Move AST implementation detail to ast_impl
